+++ example/test.dart
@@ -1,0 +1,7 @@
+
+
+import 'package:happypass/happypass.dart';
+
+void main() {
+	print(PassHttpUtils.resolveUrl("https://a.www.baidu.com:8080/asd"));
+}